requirements.txt
streamlit>=1.32.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.11.0
scikit-image>=0.22.0
